examples/teleop_to_omx/pyproject.toml
[project]
name = "teleop_android"
version = "0.1.0"
description = "Start a WebSocket server for listening for the phone's position/orientation from an Android device."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.0",
    "lerobot[feetech,phone]>=0.4.0",
    "uvicorn[standard]>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["teleop_android"]

[project.optional-dependencies]
rerun = [
    "numpy>=2.3.4",
    "rerun-sdk>=0.26.2",
    "scipy>=1.16.3",
    "transforms3d>=0.4.2",
]
